Added separate operations and improved some functions
