+++ packages/nodes-base/nodes/Aws/IAM/descriptions/group/Group.resource.ts
@@ -1,0 +1,131 @@
+import type { INodeProperties } from 'n8n-workflow';
+
+import * as create from './create.operation';
+import * as del from './delete.operation';
+import * as get from './get.operation';
+import * as getAll from './getAll.operation';
+import * as update from './update.operation';
+import { preDeleteGroup, presendGroupFields, processGroupsResponse } from '../../helpers/utils';
+import { handleError } from '../../helpers/errorHandler';
+
+export const description: INodeProperties[] = [
+	{
+		displayName: 'Operation',
+		name: 'operation',
+		type: 'options',
+		noDataExpression: true,
+		default: 'getAll',
+		displayOptions: {
+			show: {
+				resource: ['group'],
+			},
+		},
+		options: [
+			{
+				name: 'Create',
+				value: 'create',
+				description: 'Create a new group',
+				routing: {
+					send: {
+						preSend: [presendGroupFields],
+					},
+					request: {
+						method: 'POST',
+						url: '/?Action=CreateGroup&Version=2010-05-08',
+						ignoreHttpStatusErrors: true,
+					},
+					output: {
+						postReceive: [handleError],
+					},
+				},
+				action: 'Create group',
+			},
+			{
+				name: 'Delete',
+				value: 'delete',
+				description: 'Delete an existing group',
+				routing: {
+					send: {
+						preSend: [presendGroupFields, preDeleteGroup],
+					},
+					request: {
+						method: 'POST',
+						url: '/?Action=DeleteGroup&Version=2010-05-08',
+						ignoreHttpStatusErrors: true,
+					},
+					output: {
+						postReceive: [handleError],
+					},
+				},
+				action: 'Delete group',
+			},
+			{
+				name: 'Get',
+				value: 'get',
+				description: 'Retrieve details of an existing group',
+				routing: {
+					send: {
+						preSend: [presendGroupFields],
+					},
+					request: {
+						method: 'POST',
+						url: '/?Action=GetGroup&Version=2010-05-08',
+						ignoreHttpStatusErrors: true,
+					},
+					output: {
+						postReceive: [handleError, processGroupsResponse],
+					},
+				},
+				action: 'Get group',
+			},
+			{
+				name: 'Get Many',
+				value: 'getAll',
+				description: 'Retrieve a list of groups',
+				routing: {
+					send: {
+						preSend: [presendGroupFields],
+					},
+					request: {
+						method: 'POST',
+						url: '/?Action=ListGroups&Version=2010-05-08',
+						qs: {
+							pageSize:
+								'={{ $parameter["limit"] ? ($parameter["limit"] < 60 ? $parameter["limit"] : 60) : 60 }}',
+						},
+						ignoreHttpStatusErrors: true,
+					},
+					output: {
+						postReceive: [handleError, processGroupsResponse],
+					},
+				},
+				action: 'Get many groups',
+			},
+			{
+				name: 'Update',
+				value: 'update',
+				description: 'Update an existing group',
+				routing: {
+					send: {
+						preSend: [presendGroupFields],
+					},
+					request: {
+						method: 'POST',
+						url: '/?Action=UpdateGroup&Version=2010-05-08',
+						ignoreHttpStatusErrors: true,
+					},
+					output: {
+						postReceive: [handleError],
+					},
+				},
+				action: 'Update group',
+			},
+		],
+	},
+
+	...create.description,
+	...del.description,
+	...get.description,
+	...getAll.description,
+	...update.description,
+];
