+++ packages/nodes-base/nodes/Aws/IAM/descriptions/group/delete.operation.ts
@@ -1,0 +1,51 @@
+import { updateDisplayOptions, type INodeProperties } from 'n8n-workflow';
+
+const properties: INodeProperties[] = [
+	{
+		displayName: 'Group',
+		name: 'groupName',
+		default: {
+			mode: 'list',
+			value: '',
+		},
+		description: 'Select the group you want to delete',
+		modes: [
+			{
+				displayName: 'From list',
+				name: 'list',
+				type: 'list',
+				typeOptions: {
+					searchListMethod: 'searchGroups',
+					searchable: true,
+				},
+			},
+			{
+				displayName: 'By Name',
+				name: 'GroupName',
+				type: 'string',
+				hint: 'Enter the group name',
+				validation: [
+					{
+						type: 'regex',
+						properties: {
+							regex: '^[\\w+=,.@-]+$',
+							errorMessage: 'The group name must follow the allowed pattern.',
+						},
+					},
+				],
+				placeholder: 'e.g. Admins',
+			},
+		],
+		required: true,
+		type: 'resourceLocator',
+	},
+];
+
+const displayOptions = {
+	show: {
+		resource: ['group'],
+		operation: ['delete'],
+	},
+};
+
+export const description = updateDisplayOptions(displayOptions, properties);
