+++ packages/nodes-base/nodes/Aws/IAM/descriptions/group/get.operation.ts
@@ -1,0 +1,58 @@
+import { updateDisplayOptions, type INodeProperties } from 'n8n-workflow';
+
+const properties: INodeProperties[] = [
+	{
+		displayName: 'Group',
+		name: 'groupName',
+		required: true,
+		type: 'resourceLocator',
+		default: {
+			mode: 'list',
+			value: '',
+		},
+		description: 'Select the group you want to retrieve',
+		modes: [
+			{
+				displayName: 'From list',
+				name: 'list',
+				type: 'list',
+				typeOptions: {
+					searchListMethod: 'searchGroups',
+					searchable: true,
+				},
+			},
+			{
+				displayName: 'By Name',
+				name: 'GroupName',
+				type: 'string',
+				hint: 'Enter the group name',
+				validation: [
+					{
+						type: 'regex',
+						properties: {
+							regex: '^[\\w+=,.@-]+$',
+							errorMessage: 'The group name must follow the allowed pattern.',
+						},
+					},
+				],
+				placeholder: 'e.g. Admins',
+			},
+		],
+	},
+	{
+		displayName: 'Include Users',
+		name: 'includeUsers',
+		type: 'boolean',
+		default: false,
+		description: 'Whether to include a list of users in the group',
+	},
+];
+
+const displayOptions = {
+	show: {
+		resource: ['group'],
+		operation: ['get'],
+	},
+};
+
+export const description = updateDisplayOptions(displayOptions, properties);
