+++ packages/nodes-base/nodes/Aws/IAM/descriptions/index.ts
@@ -1,2 +1,2 @@
-export * from './UserDescription';
-export * from './GroupDescription';
+export * as group from './group/Group.resource';
+export * as user from './user/User.resource';
