+++ packages/nodes-base/nodes/Aws/IAM/descriptions/user/addToGroup.operation.ts
@@ -1,0 +1,89 @@
+import { updateDisplayOptions, type INodeProperties } from 'n8n-workflow';
+
+const properties: INodeProperties[] = [
+	{
+		displayName: 'User',
+		name: 'userName',
+		required: true,
+		type: 'resourceLocator',
+		default: {
+			mode: 'list',
+			value: '',
+		},
+		description: 'Select the user you want to add to the group',
+		modes: [
+			{
+				displayName: 'From list',
+				name: 'list',
+				type: 'list',
+				typeOptions: {
+					searchListMethod: 'searchUsers',
+					searchable: true,
+				},
+			},
+			{
+				displayName: 'By Name',
+				name: 'UserName',
+				type: 'string',
+				hint: 'Enter the user name',
+				validation: [
+					{
+						type: 'regex',
+						properties: {
+							regex: '^[\\w+=,.@-]+$',
+							errorMessage: 'The user name must follow the allowed pattern.',
+						},
+					},
+				],
+				placeholder: 'e.g. Admins',
+			},
+		],
+	},
+	{
+		displayName: 'Group',
+		name: 'groupName',
+		required: true,
+		type: 'resourceLocator',
+		default: {
+			mode: 'list',
+			value: '',
+		},
+		description: 'Select the group you want to add the user to',
+		modes: [
+			{
+				displayName: 'From list',
+				name: 'list',
+				type: 'list',
+				typeOptions: {
+					searchListMethod: 'searchGroups',
+					searchable: true,
+				},
+			},
+			{
+				displayName: 'By Name',
+				name: 'UserName',
+				type: 'string',
+				hint: 'Enter the group name',
+				validation: [
+					{
+						type: 'regex',
+						properties: {
+							regex: '^[\\w+=,.@-]+$',
+							errorMessage: 'The group name must follow the allowed pattern.',
+						},
+					},
+				],
+				placeholder: 'e.g. Admins',
+			},
+		],
+	},
+];
+
+const displayOptions = {
+	show: {
+		resource: ['user'],
+		operation: ['addToGroup'],
+	},
+};
+
+export const description = updateDisplayOptions(displayOptions, properties);
