+++ packages/nodes-base/nodes/Aws/IAM/descriptions/user/update.operation.ts
@@ -1,0 +1,80 @@
+import { updateDisplayOptions, type INodeProperties } from 'n8n-workflow';
+
+const properties: INodeProperties[] = [
+	{
+		displayName: 'User',
+		name: 'userName',
+		required: true,
+		type: 'resourceLocator',
+		default: {
+			mode: 'list',
+			value: '',
+		},
+		description: 'Select the user you want to update',
+		modes: [
+			{
+				displayName: 'From list',
+				name: 'list',
+				type: 'list',
+				typeOptions: {
+					searchListMethod: 'searchUsers',
+					searchable: true,
+				},
+			},
+			{
+				displayName: 'By Name',
+				name: 'UserName',
+				type: 'string',
+				hint: 'Enter the user name',
+				validation: [
+					{
+						type: 'regex',
+						properties: {
+							regex: '^[\\w+=,.@-]+$',
+							errorMessage: 'The user name must follow the allowed pattern.',
+						},
+					},
+				],
+				placeholder: 'e.g. Admins',
+			},
+		],
+	},
+	{
+		displayName: 'New User Name',
+		name: 'newUserName',
+		default: '',
+		placeholder: 'e.g. JohnSmith',
+		type: 'string',
+		validateType: 'string',
+		required: true,
+		typeOptions: {
+			regex: '^[a-zA-Z0-9+=,.@_-]+$',
+		},
+	},
+	{
+		displayName: 'Additional Fields',
+		name: 'additionalFields',
+		type: 'collection',
+		placeholder: 'Add Option',
+		default: {},
+		options: [
+			{
+				displayName: 'New Path',
+				name: 'newPath',
+				type: 'string',
+				validateType: 'string',
+				default: '/',
+				placeholder: 'e.g. /division_abc/subdivision_xyz/',
+			},
+		],
+	},
+];
+
+const displayOptions = {
+	show: {
+		resource: ['user'],
+		operation: ['update'],
+	},
+};
+
+export const description = updateDisplayOptions(displayOptions, properties);
