+++ packages/nodes-base/nodes/Aws/IAM/methods/index.ts
@@ -1,0 +1,1 @@
+export * as listSearch from './listSearch';
