+++ packages/nodes-base/nodes/Aws/IAM/methods/listSearch.ts
@@ -1,0 +1,130 @@
+import {
+	NodeApiError,
+	type IDataObject,
+	type IExecuteSingleFunctions,
+	type IHttpRequestOptions,
+	type ILoadOptionsFunctions,
+	type INodeListSearchItems,
+	type INodeListSearchResult,
+} from 'n8n-workflow';
+
+import type {
+	GetAllGroupsResponseBody,
+	GetAllUsersResponseBody,
+	GetGroupResponseBody,
+} from '../helpers/types';
+import { makeAwsRequest } from '../transport';
+
+function formatResults(items: IDataObject[], filter?: string): INodeListSearchItems[] {
+	return items
+		.map((item) => ({
+			name: String(item.UserName ?? item.GroupName ?? ''),
+			value: String(item.UserName ?? item.GroupName ?? ''),
+		}))
+		.filter(({ name }) => !filter || name.includes(filter))
+		.sort((a, b) => a.name.localeCompare(b.name));
+}
+
+export async function searchUsers(
+	this: ILoadOptionsFunctions,
+	filter?: string,
+): Promise<INodeListSearchResult> {
+	const opts: IHttpRequestOptions = {
+		method: 'POST',
+		url: '/?Action=ListUsers&Version=2010-05-08',
+	};
+
+	const responseData = (await makeAwsRequest.call(this, opts)) as GetAllUsersResponseBody;
+	const users = responseData.ListUsersResponse.ListUsersResult.Users;
+
+	if (!users) {
+		return { results: [] };
+	}
+
+	const results = formatResults(users, filter);
+
+	return {
+		results,
+	};
+}
+
+export async function searchGroups(
+	this: ILoadOptionsFunctions,
+	filter?: string,
+): Promise<INodeListSearchResult> {
+	const opts: IHttpRequestOptions = {
+		method: 'POST',
+		url: '/?Action=ListGroups&Version=2010-05-08',
+		headers: {
+			'Cache-Control': 'no-cache',
+			Pragma: 'no-cache',
+		},
+	};
+
+	const responseData = (await makeAwsRequest.call(this, opts)) as GetAllGroupsResponseBody;
+	const groups = responseData.ListGroupsResponse.ListGroupsResult.Groups;
+
+	if (!groups) {
+		return { results: [] };
+	}
+
+	const results = formatResults(groups, filter);
+
+	return {
+		results,
+	};
+}
+
+export async function searchGroupsForUser(
+	this: ILoadOptionsFunctions | IExecuteSingleFunctions,
+	filter?: string,
+	_paginationToken?: string,
+): Promise<INodeListSearchResult> {
+	const userName = (this.getNodeParameter('userName') as IDataObject)?.value;
+	if (!userName) {
+		throw new NodeApiError(
+			this.getNode(),
+			{},
+			{ message: 'User name is required to search groups.' },
+		);
+	}
+
+	const groupsData = (await makeAwsRequest.call(this, {
+		method: 'POST',
+		url: '/?Action=ListGroups&Version=2010-05-08',
+	})) as GetAllGroupsResponseBody;
+
+	const groups = groupsData.ListGroupsResponse?.ListGroupsResult?.Groups;
+
+	if (!groups || groups.length === 0) {
+		return { results: [] };
+	}
+
+	const groupCheckPromises = groups.map(async (group) => {
+		const groupName = group.GroupName as string;
+		if (!groupName) return null;
+
+		try {
+			const getGroupResponse = (await makeAwsRequest.call(this, {
+				method: 'POST',
+				url: `/?Action=GetGroup&Version=2010-05-08&GroupName=${encodeURIComponent(groupName)}`,
+			})) as GetGroupResponseBody;
+
+			const groupResult = getGroupResponse.GetGroupResponse?.GetGroupResult;
+
+			if (groupResult?.Users?.some((user) => user.UserName === userName)) {
+				return { UserName: userName, GroupName: groupName };
+			}
+		} catch (error) {}
+
+		return null;
+	});
+
+	const validUserGroups = (await Promise.all(groupCheckPromises)).filter(Boolean) as IDataObject[];
+
+	const formattedResults = formatResults(validUserGroups, filter);
+
+	return {
+		results: formattedResults,
+	};
+}
